moved settings from ViewMeasurementsActivity to PreferencesActivity
