+++ app/src/main/java/nl/nfi/cellscanner/ViewMeasurementsActivity.java
@@ -1,6 +1,7 @@
 package nl.nfi.cellscanner;
 
 import android.Manifest;
+import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -9,9 +10,11 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
+import android.icu.text.AlphabeticIndex;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
 import android.provider.Settings;
 import android.text.Html;
 import android.util.Log;
@@ -46,7 +49,7 @@ import static nl.nfi.cellscanner.recorder.RecorderUtils.gpsRecordingState;
 import static nl.nfi.cellscanner.recorder.RecorderUtils.inRecordingState;
 
 
-public class ViewMeasurementsActivity extends AppCompatActivity implements SharedPreferences.OnSharedPreferenceChangeListener {
+public class ViewMeasurementsActivity extends AppCompatActivity {
     /*
     Activity lifecycle, see: https://developer.android.com/guide/components/activities/activity-lifecycle
     Communicate Activity <-> Service ... https://www.vogella.com/tutorials/AndroidServices/article.html
@@ -56,11 +59,7 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
     private static final String TAG = ViewMeasurementsActivity.class.getSimpleName();
 
     // ui
-    private Button exportButton, clearButton;
-    private SwitchCompat swRecordingMaster, swGPSRecord, swGPSPrecision;
     private TextView vlCILastUpdate, vlGPSLastUpdate, vlGPSProvider, vlGPSLat, vlGPSLon, vlGPSAcc, vlGPSAlt, vlGPSSpeed;
-
-    private static final int PERMISSION_REQUEST_START_RECORDING = 1;
 
     /**
      * Fires when the system first creates the activity
@@ -77,21 +76,8 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
     @Override
     protected void onResume() {
         super.onResume();
+        clearGPSLocationFields();
         AppInfoActivity.showIfNoConsent(this);
-    }
-
-    @Override
-    protected void onStart() {
-        SharedPreferences sharedPreferences = this.getSharedPreferences(CellScannerApp.TITLE, MODE_PRIVATE);
-        sharedPreferences.registerOnSharedPreferenceChangeListener(this);
-        super.onStart();
-    }
-
-    @Override
-    protected void onDestroy() {
-        SharedPreferences sharedPreferences = this.getSharedPreferences(CellScannerApp.TITLE, MODE_PRIVATE);
-        sharedPreferences.unregisterOnSharedPreferenceChangeListener(this);
-        super.onDestroy();
     }
 
     private void showRecorderScreen() {
@@ -106,38 +92,6 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
         vlGPSAlt = findViewById(R.id.vlGPSAlt);
         vlGPSSpeed = findViewById(R.id.vlGPSSpeed);
 
-        exportButton = findViewById(nl.nfi.cellscanner.R.id.exportButton);
-        clearButton = findViewById(nl.nfi.cellscanner.R.id.clearButton);
-        swRecordingMaster = findViewById(nl.nfi.cellscanner.R.id.recorderSwitch);
-        swGPSRecord = findViewById(R.id.swGPSRecord);
-        swGPSPrecision = findViewById(R.id.swGPSRecordingPrecision);
-
-        /*
-         Implement checked state listener on the switch that has the ability to start or stop the recording process
-         */
-        swRecordingMaster.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                if (isChecked) requestStartRecording();
-                else RecorderUtils.stopService(getApplicationContext());
-
-            }
-        });
-
-        swGPSRecord.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                RecorderUtils.setGPSRecordingState(getApplicationContext(), isChecked);
-            }
-        });
-
-
-        swGPSPrecision.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                RecorderUtils.setGPSHighPrecisionRecordingState(getApplicationContext(), isChecked);
-            }
-        });
-
-        toggleButtonsRecordingState();
-
         // run initial update to inform the end user
         updateLogViewer();
 
@@ -150,128 +104,6 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
                     }
                 }, new IntentFilter(LocationRecordingService.LOCATION_DATA_UPDATE_BROADCAST)
         );
-    }
-
-
-    /**
-     * Request permission to the end user for Location usage. Please be aware that this request is
-     * done on a separate thread
-     */
-    private void requestLocationPermission() {
-        ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION}, PERMISSION_REQUEST_START_RECORDING);
-    }
-
-    /**
-     * Callback function for requests
-     */
-    @Override
-    public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
-        switch (requestCode) {
-            case PERMISSION_REQUEST_START_RECORDING: {
-                if (permissions.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    requestStartRecording();
-                } else {
-                    // explain the app will not be working
-                    Toast.makeText(this, "App will not work without location permissions", Toast.LENGTH_SHORT).show();
-                }
-            }
-        }
-    }
-
-
-    /**
-     * Open the terms and conditions screen on clicking the Info button
-     */
-    public void openTermsAndConditions(View view) {
-        AppInfoActivity.show(this);
-    }
-
-    /**
-     * Export data via email
-     */
-    public void exportData(View view) {
-        if (!Database.getDataFile(this).exists()) {
-            Toast.makeText(getApplicationContext(), "No database present.", Toast.LENGTH_SHORT).show();
-
-        } else {
-            String[] TO = {""};
-
-            Uri uri = FileProvider.getUriForFile(getApplicationContext(), "nl.nfi.cellscanner.fileprovider", Database.getDataFile(getApplicationContext()));
-
-            Intent sharingIntent = new Intent(android.content.Intent.ACTION_SEND);
-
-            sharingIntent.putExtra(Intent.EXTRA_EMAIL, TO);
-            sharingIntent.putExtra(android.content.Intent.EXTRA_SUBJECT, getFileTitle());
-            sharingIntent.putExtra(Intent.EXTRA_STREAM, uri);
-
-            //need this to prompts email client only
-            sharingIntent.setDataAndType(uri, "message/rfc822");
-
-            startActivity(Intent.createChooser(sharingIntent, "Share via"));
-        }
-    }
-
-
-    public void clearDatabase(View view) {
-        DialogInterface.OnClickListener dialogClickListener = new DialogInterface.OnClickListener() {
-            @Override
-            public void onClick(DialogInterface dialog, int which) {
-                Context ctx = getApplicationContext();
-                switch (which) {
-                    case DialogInterface.BUTTON_POSITIVE:
-                        CellScannerApp.resetDatabase(getApplicationContext());
-                        clearGPSLocationFields();
-                        Toast.makeText(ctx, "database deleted", Toast.LENGTH_SHORT).show();
-                        break;
-
-                    case DialogInterface.BUTTON_NEGATIVE:
-                        Toast.makeText(ctx, "Clear database cancelled", Toast.LENGTH_SHORT).show();
-                        break;
-                }
-            }
-        };
-
-        AlertDialog.Builder ab = new AlertDialog.Builder(this);
-        ab.setMessage("Drop tables. Sure?").setPositiveButton("Yes", dialogClickListener)
-                .setNegativeButton("No", dialogClickListener).show();
-    }
-
-    /**
-     * Request the start of recording user data
-     * <p>
-     * test for the right permissions, if ok, start recording. Otherwise request permissions
-     */
-    public void requestStartRecording() {
-        if (hasAccessCourseLocationPermission(this) &&
-                hasFineLocationPermission(this)) {
-            startRecording();
-        } else {
-            requestLocationPermission();
-        }
-    }
-
-    /**
-     * Start the recording service
-     */
-    private void startRecording() {
-        RecorderUtils.startService(this);
-    }
-
-    /**
-     * Set the buttons on the screen to recording state, or not recording state
-     */
-    private void toggleButtonsRecordingState() {
-        boolean isInRecordingState = inRecordingState(this);
-        swRecordingMaster.setChecked(isInRecordingState);
-
-        swGPSRecord.setEnabled(!isInRecordingState);
-        swGPSPrecision.setEnabled(!isInRecordingState);
-
-        exportButton.setEnabled(!isInRecordingState);
-        clearButton.setEnabled(!isInRecordingState);
-
-        swGPSRecord.setChecked(gpsRecordingState(this));
-        swGPSPrecision.setChecked(gpsHighPrecisionRecordingState(this));
     }
 
     // todo: Reconnect with a timer or listener, to update every x =D
@@ -312,11 +144,6 @@ public class ViewMeasurementsActivity extends AppCompatActivity implements Share
         dateFormat.setTimeZone(TimeZone.getDefault());
         Date dateTime = new Date(time);
         return dateFormat.format(dateTime);
-    }
-
-    @Override
-    public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String key) {
-        toggleButtonsRecordingState();
     }
 }
